* 'dashboard_backend' of https://git.shefcompsci.org.uk/com6103-2021-22/team12/project:
  added currency value to database, updated payment success page, added update amount of product after payment

+++ src/main/java/com/team12/foodforall/paypal/CreatePayment.java
@@ -1,20 +1,15 @@
 package com.team12.foodforall.paypal;
+
+import com.paypal.api.payments.*;
+import com.paypal.base.rest.APIContext;
+import com.paypal.base.rest.PayPalRESTException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.ArrayList;
 import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.paypal.api.payments.Amount;
-import com.paypal.api.payments.Payer;
-import com.paypal.api.payments.Payment;
-import com.paypal.api.payments.PaymentExecution;
-import com.paypal.api.payments.RedirectUrls;
-import com.paypal.api.payments.Transaction;
-import com.paypal.base.rest.APIContext;
-import com.paypal.base.rest.PayPalRESTException;
 
 @Service
 public class CreatePayment {
@@ -35,7 +30,7 @@ public class CreatePayment {
         Amount amount = new Amount();
         amount.setCurrency(currency);
         System.out.println(price);
-        double cost = new BigDecimal(Integer.parseInt(price)).setScale(2, RoundingMode.HALF_UP).doubleValue();
+        double cost = new BigDecimal(Float.parseFloat(price)).setScale(2, RoundingMode.HALF_UP).doubleValue();
         total_price = total*cost;
         amount.setTotal(String.format("%.2f", total_price));
